# src/main/resources/swaggerReservas.yaml
swagger: '2.0'
host: localhost:8080
basePath: /reservas
produces:
  - aplication/json
info:
  version: 1.0.0
  title: IES Jándula Website
  description: This is my project description
  contact:
    name: Joaquín Moreno Sánchez
    email: jmorsan426@g.educaand.es
    url: https://github.com/jmorsan/
schemes:
  - http
  - https
  
################################################################################################################
#                                              PATHS                                                           #
#############################################################################################################
paths:
  /mostrar_reservas:
    get:
      operationId: getReservas
      description: Devuelve las reservas
      responses:
        200:
          description: Devuelve todas las reservas

  /mostrar_reservas_aula:
    get:
      operationId: getReservasAula
      description: Devuelve las reservas de aula
      responses:
        200:
          description: Devuelve una lista con todas las reservas de aula
          schema:
            type: array
            items:
              $ref: '#/definitions/ReservaAula'

  /mostrar_reservas_carrito_tablets:
    get:
      operationId: getReservasCarritoTablets
      description: Devuelve las reservas de carritos tablets
      responses:
        200:
          description: Devuelve una lista con todas las reservas de carritos tablets
          schema:
            type: array
            items:
              $ref: '#/definitions/ReservaCarritoTablets'

  /mostrar_reservas_carritopcs:
    get:
      operationId: getReservasCarritoPcs
      description: Devuelve las reservas de carritos pcs
      responses:
        200:
          description: Devuelve una lista con todas las reservas de carritos pcs
          schema:
            type: array
            items:
              $ref: '#/definitions/ReservaCarritoPcs'

  
  /reservarAula:
    post:
      operationId: reservarAula
      description: reserva de un aula
      parameters:
        - name: idProfesor
          description: id del profesor
          in: query
          type: integer
          format: int64
          required: true
        - name: idAulaInformatica
          description: id del aula
          in: query
          type: integer
          format: int64
          required: true
        - name: date
          description: fecha de la reserva
          in: query
          type: integer
          format: int64
          required: true
      responses:
        200:
          description: Reserva de aula exitosa
          schema:
            $ref: '#/definitions/ReservaAula'
        404:
          description: Profesor o Aula no encontrada
        500:
          description: Error Fatal

  /reservar_carrito_pcs:
    post:
      operationId: reservarCarritoPcs
      description: reserva de un carrito de pcs
      parameters:
        - name: idProfesor
          description: id del profesor
          in: query
          type: integer
          format: int64
          required: true
        - name: idCarritoPcs
          description: id del carrito de pcs
          in: query
          type: integer
          format: int64
          required: true
        - name: date
          description: fecha de la reserva
          in: query
          type: integer
          format: int64
          required: true
        - name: ubicacionPrestamo
          description: ubicacion del carrito
          in: query
          type: string
          required: true
      responses:
        200:
          description: Reserva de aula exitosa
          schema:
            $ref: '#/definitions/ReservaCarritoPcs'
        404:
          description: Profesor o Carrito pcs no encontrado
        500:
          description: Error Fatal

  /reservar_carrito_tablets:
    post:
      operationId: reservarCarritoTablets
      description: reserva de un carrito de tablets
      parameters:
        - name: idProfesor
          description: id del profesor
          in: query
          type: integer
          format: int64
          required: true
        - name: idCarritoTablets
          description: id del carrito de tablets
          in: query
          type: integer
          format: int64
          required: true
        - name: date
          description: fecha de la reserva
          in: query
          type: integer
          format: int64
          required: true
        - name: ubicacionPrestamo
          description: ubicacion del carrito
          in: query
          type: string
          required: true
      responses:
        200:
          description: Reserva de aula exitosa
          schema:
            $ref: '#/definitions/ReservaCarritoTablets'
        404:
          description: Profesor o Carrito pcs no encontrado
        500:
          description: Error Fatal

  /confirmar:
    post:
      operationId: confirmacion
      description: Almacena los datos en el servidor
      responses:
        200:
          description: Exito al almacenar los datos en el servidor
        500:
          description: Error Fatal

  /cancelar_reserva_aula:
    delete:
      operationId: cancelarReservaAula
      description: cancela una reserva de un aula
      parameters:
        - name: idProfesor
          description: id del profesor
          in: query
          type: integer
          format: int64
          required: true
        - name: idAulaInformatica
          description: id del aula
          in: query
          type: integer
          format: int64
          required: true
        - name: date
          description: fecha de la reserva
          in: query
          type: integer
          format: int64
          required: true
      responses:
        200:
          description: Cancelacion de la reserva exitosa

        404:
          description: Reserva no encontrada en la BBDD/Sesion
        500:
          description: Error Fatal

  /cancelar_reserva_carrito_pcs:
    delete:
      operationId: cancelarReservaCarritoPcs
      description: cancela una reserva de un carrito de pcs
      parameters:
        - name: idProfesor
          description: id del profesor
          in: query
          type: integer
          format: int64
          required: true
        - name: idCarritoPcs
          description: id del carrito de pcs
          in: query
          type: integer
          format: int64
          required: true
        - name: date
          description: fecha de la reserva
          in: query
          type: integer
          format: int64
          required: true
        - name: ubicacionPrestamo
          description: ubicacion del carrito
          in: query
          type: string
          required: true
      responses:
        200:
          description: Cancelacion de la reserva exitosa

        404:
          description: Reserva no encontrada en la BBDD/Sesion
        500:
          description: Error Fatal

  /cancelar_reserva_carrito_tablets:
    delete:
      operationId: cancelarReservaCarritoTablets
      description: cancela la reserva de un carrito de tablets
      parameters:
        - name: idProfesor
          description: id del profesor
          in: query
          type: integer
          format: int64
          required: true
        - name: idCarritoTablets
          description: id del carrito de tablets
          in: query
          type: integer
          format: int64
          required: true
        - name: date
          description: fecha de la reserva
          in: query
          type: integer
          format: int64
          required: true
        - name: ubicacionPrestamo
          description: ubicacion del carrito
          in: query
          type: string
          required: true
      responses:
        200:
          description: Cancelacion de la reserva exitosa

        404:
          description: Reserva no encontrada en la BBDD/Sesion
        500:
          description: Error Fatal

definitions:
  ReservaAula:
    type: object
    description: reserva de aula
    properties:
      fecha:
        description: Fecha de la reserva
        type: number
      idProfesor:
        description: Identificador del profesor
        type: number
      idAularInformatica:
        description: Identificador del la aula de informatica
        type: number
  
  ReservaCarritoTablets:
    type: object
    description: reserva de carritos de tablets
    properties:
      fecha:
        description: Fecha de la reserva
        type: number
      idProfesor:
        description: Identificador del profesor
        type: number
      idCarritoTablets:
        description: Identificador del carrito de tablets
        type: number
        
  ReservaCarritoPcs:
    type: object
    description: reserva de carritos de pcs
    properties:
      fecha:
        description: Fecha de la reserva
        type: number
      idProfesor:
        description: Identificador del profesor
        type: number
      idCarritoPcs:
        description: Identificador del carrito de pcs
        type: number
        
  ReservasError:
    type: object
    properties:
      errorId:
        type: string
        description: This is the error id
      mensaje:
        type: string
        description: This is the mensaje of error